+++ src/plugins/data_source_management/public/index.ts
@@ -12,4 +12,5 @@ export function plugin() {
 }
 export { DataSourceManagementPluginStart } from './types';
 export { DataSourceSelector } from './components/data_source_selector';
+export { DataSourceMenu } from './components/data_source_menu';
 export { DataSourceManagementPlugin, DataSourceManagementPluginSetup } from './plugin';